Gate process sample log is added

+++ gr-rfid/include/rfid/global_vars.h
@@ -92,12 +92,12 @@ namespace gr {
     // Duration in us
     const int DR_D          = 8;
     const float BLF_D     = T_READER_FREQ / pow(10,6);  // 0.04
-    const int TPRI_D      = 1 / BLF_D;  // 25us
-    const int PW_D        = 24; // Half Tari
-    const int RTCAL_D     = 6 * PW_D; // 72us
-    const int TRCAL_D     = DR_D / BLF_D; // 200us
-    const int T1_D        = std::max(RTCAL_D, 10 * TPRI_D);  // 250us
-    const int T2_D        = 20 * TPRI_D;  // 500us
+    const float TPRI_D      = 1 / BLF_D;  // 25us
+    const float PW_D        = 24; // Half Tari
+    const float RTCAL_D     = 6 * PW_D; // 72us
+    const float TRCAL_D     = DR_D / BLF_D; // 200us
+    const float T1_D        = std::max(RTCAL_D, 10 * TPRI_D);  // 250us
+    const float T2_D        = 20 * TPRI_D;  // 500us
 
     const int CW_D         = 250;    // Carrier wave
     const int P_DOWN_D     = 2000;    // power down
